+++ cmd/clone.go
@@ -23,6 +23,7 @@ import (
 var (
 	protocol          string
 	path              string
+	parentFolder      string
 	branch            string
 	token             string
 	cloneType         string
@@ -162,8 +163,9 @@ func cloneFunc(cmd *cobra.Command, argz []string) {
 		os.Exit(1)
 	}
 
+	parseParentFolder(argz)
 	args = argz
-	targetCloneSource = argz[0]
+  targetCloneSource = argz[0]
 
 	CloneAllRepos()
 }
@@ -211,7 +213,7 @@ func getAllUserCloneUrls() ([]repo.Data, error) {
 }
 
 func createDirIfNotExist() {
-	if _, err := os.Stat(os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO") + targetCloneSource + "_ghorg"); os.IsNotExist(err) {
+  if _, err := os.Stat(os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO") + parentFolder + "_ghorg"); os.IsNotExist(err) {
 		err = os.MkdirAll(os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO"), 0700)
 		if err != nil {
 			panic(err)
@@ -354,10 +356,10 @@ func CloneAllRepos() {
 				path = repo.Path
 			}
 
-			repoDir := os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO") + targetCloneSource + "_ghorg" + "/" + path
+			repoDir := os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO") + parentFolder + "_ghorg" + "/" + path
 
 			if os.Getenv("GHORG_BACKUP") == "true" {
-				repoDir = os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO") + targetCloneSource + "_ghorg_backup" + "/" + path
+				repoDir = os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO") + parentFolder + "_ghorg_backup" + "/" + path
 			}
 
 			if repoExistsLocally(repoDir) == true {
@@ -399,6 +401,7 @@ func CloneAllRepos() {
 						return
 					}
 
+					// TODO: handle case where repo was removed, should not give user an error
 					cmd = exec.Command("git", "pull", "origin", branch)
 					cmd.Dir = repoDir
 					err = cmd.Run()
@@ -424,6 +427,19 @@ func CloneAllRepos() {
 					cloneErrors = append(cloneErrors, e)
 					return
 				}
+
+				// TODO: make configs around remote name
+				// we clone with api-key in clone url
+				args = []string{"remote", "set-url", "origin", repo.URL}
+				cmd = exec.Command("git", args...)
+				cmd.Dir = repoDir
+				err = cmd.Run()
+
+				if err != nil {
+					e := fmt.Sprintf("Problem trying to set remote on Repo: %s Error: %v", repo.URL, err)
+					cloneErrors = append(cloneErrors, e)
+					return
+				}
 			}
 
 			colorlog.PrintSuccess("Success " + repo.URL)
@@ -437,9 +453,9 @@ func CloneAllRepos() {
 
 	// TODO: fix all these if else checks with ghorg_backups
 	if os.Getenv("GHORG_BACKUP") == "true" {
-		colorlog.PrintSuccess(fmt.Sprintf("Finished! %s%s_ghorg_backup", os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO"), targetCloneSource))
+		colorlog.PrintSuccess(fmt.Sprintf("Finished! %s%s_ghorg_backup", os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO"), parentFolder))
 	} else {
-		colorlog.PrintSuccess(fmt.Sprintf("Finished! %s%s_ghorg", os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO"), targetCloneSource))
+		colorlog.PrintSuccess(fmt.Sprintf("Finished! %s%s_ghorg", os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO"), parentFolder))
 	}
 }
 
@@ -490,4 +506,8 @@ func addTokenToHTTPSCloneURL(url string, token string) string {
 	}
 
 	return "https://" + token + "@" + splitURL[1]
+}
+
+func parseParentFolder(argz []string) {
+	parentFolder = strings.ToLower(argz[0])
 }